Remove Exception from Core, to remove circular dependencies
