+++ core/index.ts
@@ -21,9 +21,7 @@ export * from './RefreshContextView';
 export * from './HoistAppModel';
 export * from './HoistAuthModel';
 
-export * from './exception/ExceptionHandler';
-export * from './exception/Exception';
-export * from './exception/Types';
+export * from './ExceptionHandler';
 
 // Explicitly exporting `XH` helps IntelliJ suggest the correct import from this core package.
 export {XH, XHApi} from './XH';
